refactor(form-styles): import output color styles from Converter helper file
